+++ src/raisedcos.cpp
@@ -5,8 +5,9 @@
 NumericVector raisedcos_(NumericVector x, NumericVector ctx)
 {
     double low = ctx[0];
-    double ctr = ctx[1];
-    double big = ctx[2];
+    double ctr1 = ctx[1];
+    double ctr2 = ctx[ctx.size() - 2];
+    double big = ctx[ctx.size() - 1];
 
     NumericVector res(x.size());
     for (int i = 0; i < x.size(); ++i) {
@@ -16,23 +17,23 @@ NumericVector raisedcos_(NumericVector x, NumericVector ctx)
             res[i] = NAN;
         } else if (x[i] < low || x[i] > big) {
             res[i] = 0;
-        } else if (x[i] < ctr) {
+        } else if (x[i] < ctr1) {
             if (low == R_NegInf) {
                 res[i] = 1;
-            } else if (low == ctr) {
+            } else if (low == ctr1) {
                 res[i] = 0;
             } else {
-                res[i] = (cos((x[i] - ctr) * M_PI / (ctr - low)) + 1) / 2;
+                res[i] = (cos((x[i] - ctr1) * M_PI / (ctr1 - low)) + 1) / 2;
             }
-        } else if (x[i] == ctr) {
+        } else if (x[i] <= ctr2) {
             res[i] = 1;
         } else {
             if (big == R_PosInf) {
                 res[i] = 1;
-            } else if (ctr == big) {
+            } else if (ctr2 == big) {
                 res[i] = 0;
             } else {
-                res[i] = (cos((x[i] - ctr) * M_PI / (big - ctr)) + 1) / 2;
+                res[i] = (cos((x[i] - ctr2) * M_PI / (big - ctr2)) + 1) / 2;
             }
         }
     }